move the fixes to fastcp.py
